Update sql result mapping

+++ src/main/java/com/example/happy_fitness/entity/Cart.java
@@ -11,8 +11,8 @@ import lombok.Data;
         classes = @ConstructorResult(
                 targetClass = com.example.happy_fitness.dto.CartDto.class,
                 columns = {
-                        @ColumnResult(name = "id", type = Float.class),
-                        @ColumnResult(name = "facilityProductId", type = Float.class),
+                        @ColumnResult(name = "id", type = Long.class),
+                        @ColumnResult(name = "facilityProductId", type = Long.class),
                         @ColumnResult(name = "code", type = String.class),
                         @ColumnResult(name = "name", type = String.class),
                         @ColumnResult(name = "image", type = String.class),
